requirements.txt
pillow==2.0.0
numpy==1.7.1